Remove wrapper markup around `Connection` children
The component only shows markup passed as children. This should
allow styling the markup in parent components without the need
to target the `Connection` component styles through global
overrides.

It also allows parent components to be blank if they decide to. For
instance, a parent component may show no markup if the connection
status is "online".